+++ src/elements/Product.php
@@ -184,8 +184,8 @@ class Product extends Purchasable
         $sources[] = ['heading' => Craft::t('digital-products', 'Product Types')];
 
         foreach ($productTypes as $productType) {
-            $key = 'productType:' . $productType->id;
-            $canEditProducts = Craft::$app->getUser()->checkPermission('digitalProducts-manageProductType:' . $productType->id);
+            $key = 'productType:' . $productType->uid;
+            $canEditProducts = Craft::$app->getUser()->checkPermission('digitalProducts-manageProductType:' . $productType->uid);
 
             $sources[$key] = [
                 'key' => $key,
@@ -349,9 +349,9 @@ class Product extends Purchasable
     public function getIsEditable(): bool
     {
         if ($this->getType()) {
-            $id = $this->getType()->id;
+            $uid = $this->getType()->uid;
 
-            return Craft::$app->getUser()->checkPermission('digitalProducts-manageProductType:' . $id);
+            return Craft::$app->getUser()->checkPermission('digitalProducts-manageProductType:' . $uid);
         }
 
         return false;